termserver/requirements.txt
Django==1.7c1
Fabric==1.9.0
Markdown==2.4.1
Pygments==1.6
amqp==1.4.5
anyjson==0.3.3
argparse==1.2.1
billiard==3.3.0.18
celery==3.1.12
coverage==3.7.1
django-extensions==1.3.8
django-filter==0.7
django-nose==1.2
django-rest-swagger==0.1.14
djangorestframework==2.3.14
djangorestframework-camel-case==0.1.3
ecdsa==0.11
ipython==2.1.0
kombu==3.0.21
nose==1.3.3
paramiko==1.14.0
pathlib==1.0
psycopg2==2.5.3
pycrypto==2.6.1
pytz==2014.4
pyzmq==14.3.1
redis==2.10.1
six==1.7.3
wsgiref==0.1.2
